SA-RF :: Add BaseTest class (#16)
