+++ tests/test_tools_qa.py
@@ -1,28 +1,16 @@
-import os
 import unittest
-from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
-from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+from utils.base_test import BaseTest
 
 BASE_URL = 'https://demoqa.com'
 
 
-class TestToolsQANavigation(unittest.TestCase):
-    def setUp(self) -> None:
-        service = Service(ChromeDriverManager().install())
-        chrome_options = webdriver.ChromeOptions()
-        chrome_options.add_argument('--window-size=1920,1080')
-        if 'CICD_RUN' in os.environ:
-            chrome_options.add_argument('--no-sandbox')
-            chrome_options.add_argument('--disable-gpu')
-            chrome_options.add_argument('--disable-dev-shm-usage')
-            chrome_options.add_argument('--headless')
-        self.driver = webdriver.Chrome(service=service, options=chrome_options)
+class TestToolsQANavigation(BaseTest):
+    def setUp(self):
+        self._setUp(BASE_URL)
         self.wait = WebDriverWait(self.driver, 8)
-        self.driver.get(BASE_URL)
 
     def test_navigate_to_toolsQA(self):
         self.assertEqual(self.driver.title, 'ToolsQA')
@@ -61,9 +49,6 @@ class TestToolsQANavigation(unittest.TestCase):
         self.driver.execute_script(f"window.scrollTo(0, {self.driver.get_window_size().get('height')})")
         books_card.click()
         self.assertEqual(self.driver.current_url, f'{BASE_URL}/books')
-
-    def tearDown(self) -> None:
-        self.driver.quit()
 
 
 if __name__ == '__main__':
