+++ tests/test_toolsqa_alerts.py
@@ -1,14 +1,11 @@
-import os
 import time
 import unittest
-from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
-from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import UnexpectedAlertPresentException, TimeoutException, NoAlertPresentException, \
     NoSuchElementException
+from utils.base_test import BaseTest
 
 BASE_URL = 'https://demoqa.com/alerts'
 INSTANT_ALERT_BUTTON_ID = 'alertButton'
@@ -19,22 +16,10 @@ CONFIRMATION_RESULT_ID = 'confirmResult'
 PROMPT_RESULT_ID = 'promptResult'
 
 
-class TestToolsQAAlerts(unittest.TestCase):
+class TestToolsQAAlerts(BaseTest):
     def setUp(self):
-        service = Service(ChromeDriverManager().install())
-        chrome_options = webdriver.ChromeOptions()
-        chrome_options.add_argument('--window-size=1920,1080')
-        if 'CICD_RUN' in os.environ:
-            chrome_options.add_argument('--no-sandbox')
-            chrome_options.add_argument('--disable-gpu')
-            chrome_options.add_argument('--disable-dev-shm-usage')
-            chrome_options.add_argument('--headless')
-        self.driver = webdriver.Chrome(service=service, options=chrome_options)
+        self._setUp(BASE_URL)
         self.wait = WebDriverWait(self.driver, 10)
-        self.driver.get(BASE_URL)
-
-    def tearDown(self):
-        self.driver.quit()
 
     def test_alert(self):
         self.driver.find_element(By.ID, INSTANT_ALERT_BUTTON_ID).click()
